refactor: chakra-theme에서 chakraTheme으로 네이밍 변경
